Eventro controller e fix eventoservice

+++ src/main/java/vincenzomanfredi/EsameBE6/services/EventoService.java
@@ -8,6 +8,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import vincenzomanfredi.EsameBE6.entities.Evento;
 import vincenzomanfredi.EsameBE6.entities.Utente;
+import vincenzomanfredi.EsameBE6.exceptions.BadRequestException;
 import vincenzomanfredi.EsameBE6.exceptions.NotFoundException;
 import vincenzomanfredi.EsameBE6.payloads.EventoDTO;
 import vincenzomanfredi.EsameBE6.repositories.EventoRepository;
@@ -25,9 +26,7 @@ public class EventoService {
     }
 
     //SAVE
-    public Evento save(EventoDTO payload) {
-
-        Utente organizzatore = this.utenteService.findById(payload.organizzatoreId());
+    public Evento save(EventoDTO payload, Utente organizzatore) {
 
         Evento nuovoEvento = new Evento(
                 payload.titolo(),
@@ -59,9 +58,31 @@ public class EventoService {
                 .orElseThrow(() -> new NotFoundException("L'evento con ID " + eventoId + " non è stato trovato!"));
     }
 
-    //DELETE
-    public void findByIdAndDelete(long eventoId) {
+    //UPDATE
+    public Evento findByIdAndUpdate(long eventoId, EventoDTO payload, Utente organizzatore) {
         Evento found = this.findById(eventoId);
+
+        if (found.getOrganizzatore().getId() != organizzatore.getId()) {
+            throw new BadRequestException("Non sei autorizzato a modificare questo evento, non sei l'organizzatore!");
+        }
+
+        found.setTitolo(payload.titolo());
+        found.setDescrizione(payload.descrizione());
+        found.setData(payload.data());
+        found.setLuogo(payload.luogo());
+        found.setPosti(payload.postiDisponibili());
+
+        return this.eventoRepository.save(found);
+    }
+
+    //DELETE
+    public void findByIdAndDelete(long eventoId, Utente organizzatore) {
+        Evento found = this.findById(eventoId);
+
+        if (found.getOrganizzatore().getId() != organizzatore.getId()) {
+            throw new BadRequestException("Non sei autorizzato a eliminare questo evento, non sei l'organizzatore!");
+        }
+
         this.eventoRepository.delete(found);
         log.info("Evento con ID " + eventoId + " eliminato.");
     }
